Refactor terminal and plot
