+++ examples/LandmarkDetection/DQN/GUI/text/quick_automatic_text.html
@@ -2,6 +2,5 @@
 <p style='color:#006EAF'><i>For more information: Ctrl+H</i></p>
 <br /><b>1. Load data:</b> choose default or browse custom data.
 <br /><b>2. Task selection: </b>select (Play|Evaluation) task.
-<br /><b>3. Simulation: </b>start visualisation by pressing <i>Start</i> button, <i>Pause</i> to pause simulation
-    , <i>Terminate</i> to terminate simulation. After termination, you can run a new simulation using different settings.
+<br /><b>3. Simulation:</b> control simulation by pressing <i>Start, Pause,</i> or <i>Terminate</i>. After termination, you can run a new simulation using different settings.
 <br /><b>4. Agent speed:</b> optionally change simulation speed using agent speed slider.
